requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
